requirements.txt
Django==1.7.1
Pillow==2.6.1
Werkzeug==0.9.6
django-appconf==0.6
django-compressor==1.4
django-extensions==1.4.5
django-suit==0.2.11
django-suit-redactor==0.0.2
numpy==1.9.0
psycopg2==2.5.4
six==1.8.0
sorl-thumbnail==11.12.1b
wsgiref==0.1.2
